# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Family Growth API
  version: 0.3.0
  description: Stage 3 生产认证、家庭成长、任务提交审核、三奖励与 Wallet/Ledger API。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
  /auth/bootstrap:
    post:
      operationId: bootstrapFamily
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapRequest'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/Conflict'
  /auth/login:
    post:
      operationId: loginParent
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: PIN temporarily locked after repeated failures
  /auth/child-sessions:
    post:
      operationId: createChildSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [childId]
              properties:
                childId:
                  type: string
                  format: uuid
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /families/{familyId}/children:
    post:
      operationId: addChild
      parameters:
        - $ref: '#/components/parameters/FamilyId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChild'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /families/{familyId}/children/{childId}/plans:
    post:
      operationId: createGrowthPlan
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - $ref: '#/components/parameters/ChildId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlan'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /families/{familyId}/plans/{planId}/goals:
    post:
      operationId: createGrowthGoal
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: planId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoal'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /families/{familyId}/goals/{goalId}/tasks:
    post:
      operationId: createGrowthTask
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: goalId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTask'
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /families/{familyId}/children/{childId}/tasks/{taskId}/completions:
    post:
      operationId: submitTaskCompletion
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - $ref: '#/components/parameters/ChildId'
        - name: taskId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                evidenceNote:
                  type: [string, 'null']
                  maxLength: 1000
      responses:
        '201':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /families/{familyId}/completions/{completionId}/review:
    post:
      operationId: reviewTaskCompletion
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - name: completionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCompletion'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
  /families/{familyId}/children/{childId}/wallet:
    get:
      operationId: getWallet
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - $ref: '#/components/parameters/ChildId'
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /families/{familyId}/children/{childId}/ledger:
    get:
      operationId: getLedger
      parameters:
        - $ref: '#/components/parameters/FamilyId'
        - $ref: '#/components/parameters/ChildId'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          $ref: '#/components/responses/Ok'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    FamilyId:
      name: familyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ChildId:
      name: childId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 100
  responses:
    Ok:
      description: Operation succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Created:
      description: Resource created
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Unauthorized:
      description: Missing, invalid or expired Bearer session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Forbidden:
      description: Authenticated role cannot perform this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    NotFound:
      description: Resource does not exist or is not visible in this family scope
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
    Conflict:
      description: State or idempotency conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
  schemas:
    BootstrapRequest:
      type: object
      required: [familyName, parentName, pin]
      properties:
        familyName:
          type: string
          minLength: 1
          maxLength: 100
        parentName:
          type: string
          minLength: 1
          maxLength: 80
        pin:
          type: string
          pattern: '^\\d{6}$'
          writeOnly: true
    LoginRequest:
      type: object
      required: [familyId, parentId, pin]
      properties:
        familyId:
          type: string
          format: uuid
        parentId:
          type: string
          format: uuid
        pin:
          type: string
          pattern: '^\\d{6}$'
          writeOnly: true
    CreateChild:
      type: object
      required: [displayName, birthDate, ageStage]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 80
        birthDate:
          type: string
          format: date
        ageStage:
          type: string
          enum: [INFANT_0_2, EARLY_CHILDHOOD_2_5, CHILD_6_9, PRETEEN_10_12, TEEN_13_PLUS]
    CreatePlan:
      type: object
      required: [title, startDate]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: [string, 'null']
          maxLength: 1000
        startDate:
          type: string
          format: date
        endDate:
          type: [string, 'null']
          format: date
    CreateGoal:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: [string, 'null']
          maxLength: 1000
    CreateTask:
      type: object
      required: [title, category, difficulty, expectedMinutes]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: [string, 'null']
          maxLength: 1000
        category:
          type: string
          enum: [LEARNING, READING, EXERCISE, HABIT, CHORE, SELF_CARE, PARENT_CHILD]
        difficulty:
          type: string
          enum: [EASY, NORMAL, HARD]
        expectedMinutes:
          type: integer
          minimum: 1
          maximum: 480
    ReviewCompletion:
      type: object
      required: [approved, xpReward, coinReward, moneyReward]
      properties:
        approved:
          type: boolean
        xpReward:
          type: integer
          format: int64
          minimum: 0
        coinReward:
          type: integer
          format: int64
          minimum: 0
        moneyReward:
          type: number
          multipleOf: 0.01
          minimum: 0
        reviewNote:
          type: [string, 'null']
          maxLength: 1000
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
    ApiResponse:
      type: object
      required: [data, error, traceId]
      properties:
        data: {}
        error:
          oneOf:
            - $ref: '#/components/schemas/ApiError'
            - type: 'null'
        traceId:
          type: [string, 'null']