# family-growth-web/src/main/java/com/familygrowth/web/FamilyGrowthController.java
package com.familygrowth.web;
import com.familygrowth.application.FamilyGrowthService; import com.familygrowth.application.Stage3Service; import com.familygrowth.domain.*; import com.familygrowth.domain.Stage3Models.Actor; import jakarta.validation.Valid; import jakarta.validation.constraints.*; import java.time.LocalDate; import java.util.UUID; import org.springframework.http.*; import org.springframework.web.bind.annotation.*;
@RestController @RequestMapping("/api/v1")
public class FamilyGrowthController {
 private final FamilyGrowthService service; private final Stage3Service stage3; public FamilyGrowthController(FamilyGrowthService service,Stage3Service stage3){this.service=service;this.stage3=stage3;}
 @PostMapping("/families/{familyId}/children") @ResponseStatus(HttpStatus.CREATED) ApiResponse<ChildProfile> child(@RequestAttribute(Stage3Models.ACTOR_REQUEST_ATTRIBUTE) Actor actor,@PathVariable UUID familyId,@Valid @RequestBody CreateChild r){return ApiResponse.ok(stage3.addChild(actor,familyId,r.displayName(),r.birthDate(),r.ageStage()));}
 @PostMapping("/families/{familyId}/children/{childId}/plans") @ResponseStatus(HttpStatus.CREATED) ApiResponse<GrowthPlan> plan(@RequestAttribute(Stage3Models.ACTOR_REQUEST_ATTRIBUTE) Actor actor,@PathVariable UUID familyId,@PathVariable UUID childId,@Valid @RequestBody CreatePlan r){stage3.requireParent(actor,familyId);return ApiResponse.ok(service.createPlan(familyId,childId,r.title(),r.description(),r.startDate(),r.endDate()));}
 @PostMapping("/families/{familyId}/plans/{planId}/goals") @ResponseStatus(HttpStatus.CREATED) ApiResponse<GrowthGoal> goal(@RequestAttribute(Stage3Models.ACTOR_REQUEST_ATTRIBUTE) Actor actor,@PathVariable UUID familyId,@PathVariable UUID planId,@Valid @RequestBody CreateGoal r){stage3.requireParent(actor,familyId);return ApiResponse.ok(service.createGoal(familyId,planId,r.title(),r.description()));}
 @PostMapping("/families/{familyId}/goals/{goalId}/tasks") @ResponseStatus(HttpStatus.CREATED) ApiResponse<GrowthTask> task(@RequestAttribute(Stage3Models.ACTOR_REQUEST_ATTRIBUTE) Actor actor,@PathVariable UUID familyId,@PathVariable UUID goalId,@Valid @RequestBody CreateTask r){stage3.requireParent(actor,familyId);return ApiResponse.ok(service.createTask(familyId,goalId,r.title(),r.description(),r.category(),r.difficulty(),r.expectedMinutes()));}
 public record CreateChild(@NotBlank @Size(max=80) String displayName,@NotNull @PastOrPresent LocalDate birthDate,@NotNull AgeStage ageStage){}
 public record CreatePlan(@NotBlank @Size(max=120) String title,@Size(max=1000) String description,@NotNull LocalDate startDate,LocalDate endDate){}
 public record CreateGoal(@NotBlank @Size(max=120) String title,@Size(max=1000) String description){}
 public record CreateTask(@NotBlank @Size(max=120) String title,@Size(max=1000) String description,@NotNull TaskCategory category,@NotNull TaskDifficulty difficulty,@Min(1) @Max(480) int expectedMinutes){}
}